+++ bin-proto/src/enum_ext.rs
@@ -1,7 +1,7 @@
 use crate::Protocol;
 
 /// A trait for `enum`s with the `Protocol` trait.
-pub trait Enum: Protocol {
+pub trait EnumExt: Protocol {
     /// The type used to store the enum discriminant
     type Discriminant: Protocol;