+++ bin-proto/src/types/vec.rs
@@ -3,7 +3,7 @@ use core::any::Any;
 
 impl<T: Protocol> FlexibleArrayMember for Vec<T> {
     fn read(read: &mut dyn BitRead, settings: &Settings, ctx: &mut dyn Any) -> Result<Self, Error> {
-        util::read_list_to_eof(read, settings, ctx)
+        util::read_items_to_eof(read, settings, ctx)
     }
 
     fn write(
@@ -12,6 +12,6 @@ impl<T: Protocol> FlexibleArrayMember for Vec<T> {
         settings: &Settings,
         ctx: &mut dyn Any,
     ) -> Result<(), Error> {
-        util::write_list(self.iter(), write, settings, ctx)
+        util::write_items(self.iter(), write, settings, ctx)
     }
 }